tasks/dopetoken/app/requirements.txt
aiohttp
aiohttp_jinja2
jinja2
pyjwt
